Moving to a container and presentational pattern
