+++ frontend/src/components/recommendations/recommendationHeader.tsx
@@ -1,0 +1,10 @@
+export default function RecommendationHeader() {
+    return (
+        <section className="species-container">
+            <div className="hero species-hero">
+                <h1>Agroforestry Recommendations</h1>
+                <p>Analyze and discover optimal species for your farm</p>
+            </div>
+        </section>
+    );
+}
